Update code for breaking changes in go-github 21.0.0

+++ provider/github/review.go
@@ -117,7 +117,7 @@ func getPostedComment(ctx context.Context, client *Client, owner, repo string, n
 		}
 
 		for {
-			comments, resp, err := client.PullRequests.ListReviewComments(ctx, owner, repo, int64(number), review.GetID(), listCommentsOpts)
+			comments, resp, err := client.PullRequests.ListReviewComments(ctx, owner, repo, number, review.GetID(), listCommentsOpts)
 			if handleAPIError(resp, err) != nil {
 				return nil, err
 			}
